+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Makefile` to auto-generate project files
 - Added experimental sunodo validator mode
 - Added inspect command to `cartesi-rollups-cli`
+- Added increase-time command to `cartesi-rollups-cli`
 
 ### Changed